Begin work on hook scripts.
The only one close to done is the pre-commit hook, but once I have that, moving
forward will be easier.

Also includes a pre-commit script used by sdboyer for tracking something related
to themes (I'm not exactly sure what it was, but it was narrower in scope than
the full versioncontrol api).

Signed-off-by: Dan Hackney <chrono325@gmail.com>